LayoutTests/imported/w3c/web-platform-tests/XMLHttpRequest/open-url-worker-origin-expected.txt

Blocked access to external URL http://www2.localhost:8800/XMLHttpRequest/resources/inspect-headers.py?filter_name=origin&cors

PASS XMLHttpRequest: worker scripts, origin and referrer 
FAIL Origin header assert_equals: expected "Origin: http://localhost:8800\n" but got ""
PASS Referer header 
PASS Request URL test 
